I2C fixes

+++ include/hkk/utils/utils.hpp
@@ -11,7 +11,7 @@ constexpr uint8 lsb(const uint16 cmd) {
     return static_cast<uint8>(static_cast<uint16>(cmd) & 0xFF);
 }
 
-void sleep_us(uint32 us);
-void sleep_ms(uint32 ms);
+void sleep_us(uint64 us);
+void sleep_ms(uint64 ms);
 
 };
