+++ src/platform/rp2350/utils/utils.cpp
@@ -5,7 +5,7 @@
 
 namespace hkk::utils {
 
-inline void sleep_us(uint32 us) {::sleep_us(us);}
-inline void sleep_ms(uint32 ms) {::sleep_us(ms);}
+inline void sleep_us(uint64 us) {::sleep_us(us);}
+inline void sleep_ms(uint64 ms) {::sleep_us(ms);}
 
 }
